Change existing plugins to accessories
