+++ plugins/column/expandable-text/src/index.js
@@ -1,5 +1,5 @@
-import { OuterbasePluginCell_$PLUGIN_ID } from './views/cell.js';
+import { OuterbasePluginCellAccessory_$PLUGIN_ID } from './views/cell-accessory.js';
 import { OuterbasePluginEditor_$PLUGIN_ID } from './views/editor.js';
 
-window.customElements.define('outerbase-plugin-cell-$PLUGIN_ID', OuterbasePluginCell_$PLUGIN_ID);
+window.customElements.define('outerbase-plugin-cell-accessory-$PLUGIN_ID', OuterbasePluginCellAccessory_$PLUGIN_ID);
 window.customElements.define('outerbase-plugin-editor-$PLUGIN_ID', OuterbasePluginEditor_$PLUGIN_ID);
